pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thrml-obamify"
version = "0.1.0"
description = "Energy-based image palette stylization in the spirit of 'Obamify', implemented with thrml + JAX."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "thrml-Obamify contributors"}]
keywords = ["jax", "thrml", "energy-based", "image stylization", "palette"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "thrml==0.1.3",
  "numpy==2.1.3",
  "Pillow==10.4.0",
  "matplotlib==3.9.2",
  "imageio==2.36.0",
  "tqdm==4.66.5",
]

[project.optional-dependencies]
# Platform-specific JAX installs
cpu = [
  "jax==0.4.33",
  "jaxlib==0.4.33",
]
metal = [
  "jax>=0.4.33",
  "jax-metal>=0.0.8",
]

[project.scripts]
obamify = "thrml_obamify.cli:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
ignore = ["E203"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
